as is delta_ns
git-svn-id: https://svn.code.sf.net/p/gnetraid/code@128 e6afed3b-5b75-4bb5-bfa7-be4a7fb757eb

+++ Net-OnlineCode/trunk/clib/xor_test.c
@@ -443,7 +443,7 @@ void do_benchmarks(void) {
   // we need better than the 1s granularity provided by time()
   struct timespec start_time, end_time;
 
-  long delta_ns;
+  long long delta_ns;
   int retval;
 
   long long ns_per_test = 2500000000l; // 2.5 seconds
